requirements.txt
albumentations==1.2.0
numpy==1.20.3
opencv-python==4.6.0.66
PyYAML==6.0
scikit-learn==1.0.2
tensorboard==2.7.0
tensorboardX==2.4.1
torch==1.10.1
torchvision==0.11.2
tqdm==4.62.3
